Add `lens` entity and recipe for `detonator` (#1876)
This PR adds:

- A new entity called `lens`
    - Recipe: 1 `glass` + 1 `mithril` with a required `drill` (idea: you grind the piece of glass using a mithril-coated drill bit)
    - For now it does not convey any capabilities, but it seems ripe for some sort of pun on lenses.
- A recipe for `detonator`:
    - 1 `string` + 10 `curry` + 1 `lens`
    - Idea: curry is explosive, so you attach a string as a fuse and then ignite it by focusing sunlight with the lens.

Because of the `mithril`, these are not so easy to craft (Level 13 and 14, respectively) but I don't know any particular reason that these need to be easy to craft.  If we wanted, we could easily remove the `mithril`, which would result in them being Level 6 and 7 instead.

Closes #1017.