Restructure hyperloop definition (#1913)
Closes #1910 

I'd argue that the code is behaving correctly and as intended. Snippets wrapped with backticks are/should be treated as regular texts with highlighting. Introducing line breaks inside a snippet should trigger add a newline character to the text before being rendered (which is what was happening in the first place).

This PR readjusts the description a bit to make sure that the snippets are in one line and do not have any line breaks in between.