minimize some To/FromJSON instances (#1908)
Towards #1907.  For the most part, eschew the default To/FromJSON deriving in favor of `sumEncoding = ObjectWithSingleField`, which encodes `Constructor val` as something like `Constructor: val` instead of `{tag: Constructor, contents: val}`.  Also omit empty `Comments` records.

In the context of #1907, this reduced the size of the generated JSON by about half.