Notification popups (#2027)
Closes #916.

Adds animated popup boxes at the top of the screen to alert the user when they (1) attain a new achievement, (2) unlock new recipes, or (3) unlock new commands.

I've tried hard to make sure that the UI always gets redrawn when there is a popup box that needs to be animated, but it was a bit fiddly and it's possible I've missed a case.  Note that we used to use `continueWithoutRedraw` extensively while displaying menus, but in case a notification popup occurs we don't want that.  Rather than plumb around a boolean so we can decide whether to `continueWithoutRedraw` or not, I decided to just get rid of them.  This means we now redraw the menu every frame but that really shouldn't be an issue.

Open to suggestions for making the popups look nicer or including more information.

To test, try *e.g.* 
- Opening the About menu (make sure to delete `~/.local/share/swarm/achievement/LookedAtAboutScreen` first)
- Completing a tutorial (delete `~/.local/share/swarm/achievement/CompletedSingleTutorial` first)
- Start a Classic game and get a `tree` (should notify of new recipes)
- `make "branch"; make "branch predictor"` (should notify of new commands)