ensure non-overlapping mask char (#2079)
We want to ensure that the character used for the `mask` does not also occur as an entry in the `palette`.

# Testing
Added new test:
```
scripts/play.sh -i data/scenarios/Testing/_Validation/2077-mask-char.yaml
```
# Also in this PR

To simplify the test, we make use of the fact that `mask` is now also supported at the toplevel in addition to within a `structure` definition.