Render contiguous boundaries (#1285)
Closes #1271

## Demo

    scripts/play.sh --scenario data/scenarios/Testing/1271-wall-boundaries.yaml --autoplay

![Screenshot from 2024-09-18 09-27-14](https://github.com/user-attachments/assets/5e4ac74f-3f8c-487a-97d2-78e2320f93bb)

## Implementation

This implements neighbor-aware boundary rendering strictly on the display layer.  Underlying entities remain unchanged (for the purpose of world cell contents and e.g. `scan`).  Eventually we should be able to retire all of the niche boundary glyph entities.

Only a single function `getBoundaryDisplay` is exposed for computing the glyph appropriate for adjacent boundaries.  It takes a function from `AbsoluteDir -> Bool` indicating whether the neighbor cell in a particular direction is (also) a "boundary" entity.

The internal `glyphForNeighbors` function is implemented with pattern-matching such that exhaustive coverage of cases is guaranteed.