Refactor: move pretty-printing and failure code to `swarm-util` (#2155)
`Swarm.Language.Pretty` (in the `swarm-lang` package) had a bunch of generic pretty-printing infrastructure along with a bunch of specific instances for terms and related things.  However, the generic pretty-printing code infrastructure is not specific to swarm-lang.  This PR:

- Moves the generic pretty-printing infrastructure + utilities to `Swarm.Pretty` in `swarm-util` to make it more generally applicable.
- Distributes all the actual `PrettyPrec` instances to live next to the declarations of the corresponding data types.

In addition, `Swarm.Game.Failure` (in the `swarm-scenario` package) also had no dependencies on anything else in `swarm-scenario`, so this PR also moves it to `Swarm.Failure` in `swarm-util`, to make it more generally applicable.

Along the way I also removed `DelayType` and `Prec` since they are no longer being used.