Fix type inference for recursive let bindings (#2187)
Fixes #2186.

When inferring the type of a recursive let binding with no provided type signature, we generate a unification variable to stand for its type, then infer the type of the body with the unification variable in the context.  Finally, we must unify the generated unification variable with the inferred type.  However, previously, we were doing this final unification with a direct call to `=:=` (which simply returns an `Either`) and discarding the result, which meant that we would simply ignore it when they failed to unify.

This PR replaces the bad call to `=:=` with a call to `unify`, and also makes the `=:=` import qualified to make it less likely that we will ever call `=:=` directly like this again.

This would only have ever caused a problem with (1) a recursive function definition (2) with no type annotation (3) with a type error (4) specifically in one of the recursive applications.  Apparently no one had ever done all of 1-4 at the same time before; @xsebek was the lucky winner. :smile: