Support a generic "original" payload in structure recognizer (#2216)
The recognizer algorithm operates on a `NonEmptyGrid` of `Maybe` elements.  However, there may exist some progenitor object with auxiliary metadata payload that is independent of the algorithm but should be presented along with the final recognized structure.

So long as an "extraction" function is provided that derives the nonempty grid of `a`s from the original object `b`, we can send `b` along for the ride through the algorithm without concern to `b`'s internals.

# Changes

Overall this refactoring makes the code more generic and simpler.

* Build the recognizer automata at scenario parse time
* Enforce grid no emptiness earlier in recognizer construction
* Subsume redundant `_structureDefs` member of `StaticStructureInfo` record with `_staticAutomatons` field
* Make some type signatures in the UI more generic
* Push the conversion from `Cell` to `Entity `further up the pipeline
* Remove some derived `Show` instances (`show`-ing a Scenario is not likely to be useful due to sheer volume of content)