Make custom entities override built-in entities (#2241)
Fixes #2239.  The comments on the `Semigroup` instance for `EntityMap` *say* "it is possible for the latter 'EntityMap' to override members of the former with the same name.  This replacement happens automatically with 'Map'...", however, this was not correct.  The `Semigroup` instance for `Map` calls `Map.union`, which is *left*-biased (https://hackage.haskell.org/package/containers-0.7/docs/Data-Map-Lazy.html#t:Map).  So we want `n2 <> n1` here instead of `n1 <> n2`, so that the second `EntityMap` properly overrides the first.

Note that `c1 <> c2` is not correct either, for a more subtle reason.  In the case of overrides it needs to be manually adjusted, in a similar way to what is already done for `d1`; see #2240.  I will address that in a separate PR.

The attached scenario will succeed only if the `rock` entity is properly overridden so that it can provide the `move` capability.

Also had to apply a couple minor fixes to the `pied-piper` scenario, which it turns out tried to define custom entities named `key` and `wall` which were silently being ignored in preference to the standard entities of the same name; but then the solution inadvertently ended up depending on a property of the standard `key` (namely `pickable`) which the custom one did not have.