Make `halt`ed robots immediately wake up (#2254)
The code for `halt` used to (1) cancel the robot's CESK machine, and then (2) put it to sleep forever.  However, at some point the way we cancel CESK machines changed: instead of immediately putting the machine into a finished state, we simply put it into a state which will propagate an uncatchable error all the way up the stack.  This is necessary to make sure we finalize things appropriately, persist any definitions in the environment which ought to be persisted, etc., but it means that a "canceled" robot still has to run for a bit longer to do the finalization work.  Robots which were inactive when they were canceled thus never got a chance to finalize.  The solution is to *wake up* the robot after canceling rather than putting it to sleep!

Closes #2253.