only meet interactive robots (#2262)
As I was trying to solve #2249, I was rather confused after trying to `meet` a built robot and `give` them an item, but the item was not delivered.  It turns out that I had been `meet`-ing the `judge` robot instead.

We should distinguish between robots that are intended to be interactive or not, and only allow `meet`-ing an interactive robot.