Rename `return` to `pure` (#2285)
Closes #700.

The most interesting changes are in `swarm-lang/Swarm/Language/Parser/{Lex,Term}.hs`, where I remove the code for parsing swarm-0.5 and add an option to parse swarm-0.6 with `return` instead of `pure`, so `swarm format --v0.6` can be used to auto-convert code using `return` to code using `pure`.

Let me know how you feel about this change!  Another possible choice would be to leave `return` as a synonym for `pure`.  I could be convinced to do that as well.