Throw an error if any world description fails to parse or typecheck (#2306)
Fixes #2304.  Previously, any `data/worlds/*.world` which failed to parse or typecheck only generated a warning, which could be viewed in Messages, but could not be seen if loading a single scenario.

There is not much use to only display a warning and keep going in such a circumstance; presumably the only time this would happen is if someone is developing a world description, in which case they want to know immediately about any errors.  This PR makes such world loading failures into an exception rather than a warning.